Add missing save-exursion to semanticdb-refresh-table
(semanticdb-table::semanticdb-refresh-table): The change in rev. 8617
turned up a bug in this method.  Since the current table now correctly
has a buffer set, the first clause in the 'cond' would be taken, but
there was a `save-excursion' missing, which in turn triggered a
failure in the integration test because it would search for a symbol
in the wrong buffer where GNU global was not active.